+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased: v1.8.2
+- GCGI-1484: Refactor the `plugins/wgts/common` directory, moving code to better locations
+- GCGI-1512: Clean up `lib/djerba/data`, moving files to specific plugins where possible, `lib/djerba/util/data` otherwise
+- GCGI-1515: Make `mrdetect_filter_only` inputs optional for non-clinical reports
+- Fix issue with benchmark plugin; test no longer breaks when Djerba core version is changed
+
 ## v1.8.1: 2025-03-04
 - GCGI-1455: New `tar.status` plugin to add a display box for ctDNA status
 - GCGI-1517: Remove obsolete R scripts and tests